Fixes
More:
+ Remove unnecessary boxing in Mutable classes.
+ Remove unnecessary unboxing in MutableBoolean

+++ src/main/java/org/apache/commons/lang3/mutable/MutableByte.java
@@ -81,7 +81,7 @@ public class MutableByte extends Number implements Comparable<MutableByte>, Muta
      */
     @Override
     public Byte getValue() {
-        return Byte.valueOf(this.value);
+        return this.value;
     }
 
     /**
@@ -327,7 +327,7 @@ public class MutableByte extends Number implements Comparable<MutableByte>, Muta
      * @return a Byte instance containing the value from this mutable
      */
     public Byte toByte() {
-        return Byte.valueOf(byteValue());
+        return byteValue();
     }
 
     /**
