+++ src/main/java/org/apache/commons/lang3/mutable/MutableShort.java
@@ -81,7 +81,7 @@ public class MutableShort extends Number implements Comparable<MutableShort>, Mu
      */
     @Override
     public Short getValue() {
-        return Short.valueOf(this.value);
+        return this.value;
     }
 
     /**
@@ -327,7 +327,7 @@ public class MutableShort extends Number implements Comparable<MutableShort>, Mu
      * @return a Short instance containing the value from this mutable, never null
      */
     public Short toShort() {
-        return Short.valueOf(shortValue());
+        return shortValue();
     }
 
     /**
